third_party/WebKit/LayoutTests/platform/win/webaudio/Oscillator/oscillator-basic-expected.txt

CONSOLE WARNING: line 53: The provided value '0' is not a valid enum value of type OscillatorType.
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = 1 duplicate test name: "osc.type is equal to "custom"."
PASS osc.type = 'sine' is equal to "sine". 
PASS osc.type = 'square' is equal to "square". 
PASS osc.type = 'sawtooth' is equal to "sawtooth". 
PASS osc.type = 'triangle' is equal to "triangle". 
PASS osc.type = 'custom' threw InvalidStateError: Failed to set the 'type' property on 'OscillatorNode': 'type' cannot be set directly to 'custom'.  Use setPeriodicWave() to create a custom Oscillator type.. 
PASS osc.setPeriodicWave(wave) did not throw an exception. 
PASS osc.type is equal to "custom". 
PASS osc.type = 0 is not equal to 0. 
PASS osc.type is equal to "custom". 
Harness: the test ran to completion.
